Fix oversight from #12185 where two input parsing functions were not migrated from transpile to generate_preset_pm  with the others.
